Reparato quello :)
Could be backported, but i don't care for these old code snippets.

+++ libraries/circ_buf/circ_buf.cpp
@@ -78,7 +78,7 @@ int Circ_buf::cb_is_full() {
 // cb_info() debugging help
 void Circ_buf::cb_info() const {
   std::cout << "\nBuffer:\t\t";
-  std::cout << (int) _buf;
+  std::cout << (long) _buf;
 
   std::cout << "\n  size:\t\t";
   std::cout << _size;
